back/api-back/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
dotenv
pymysql
mysql-connector-python
patentes-vehiculares-chile